Dry-up the contrib detail and contrib list views
